Lobby Refreshing(Replaces Player Refreshing)
This refreshes the entire lobby rather than just the player count so if
the host changes map it updates and if he changes gamemode it updates.

Also had to rename two gametype image files but they still work
accordingly because it uses variantType? and that gives me an acronym
for king of the hill and capture the flag but everything still works
great and I don't see any errors.